requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
dnspython==2.4.2
doc2docx==0.2.4
docx2pdf==0.1.8
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fire==0.5.0
Flask==3.0.0
Flask-PyMongo==2.3.0
fonttools==4.44.3
frozenlist==1.4.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
JPype1==1.4.1
jupyter_client==8.6.0
jupyter_core==5.5.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.1
openai==0.28.1
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pdf2docx==0.5.6
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pylokit==0.8.1
pymongo==4.6.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
pyzmq==25.1.1
reportlab==4.0.7
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tabula-py==2.8.2
termcolor==2.3.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
yarl==1.9.2
